fix/refactor: move github auth to AuthController
