+++ src/main/java/controly/controller/AuthController.java
@@ -1,12 +1,18 @@
 package controly.controller;
 
 
+import controly.config.Constant;
+import controly.dto.DataGithubPostRequest;
+import controly.dto.GithubUserRequest;
 import controly.dto.LoginRequest;
 import controly.entities.UserEntity;
 import controly.service.AuthService;
+import controly.service.GithubService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 @CrossOrigin(origins = "*")
@@ -14,9 +20,25 @@ import org.springframework.web.bind.annotation.*;
 public class AuthController {
     @Autowired
     AuthService authService;
+    @Autowired
+    private GithubService githubService;
 
     @PostMapping()
     public ResponseEntity<UserEntity> login(@RequestBody LoginRequest login){
         return authService.login(login.getEmail(), login.getSenha());
+    }
+
+    @GetMapping("/github")
+    public ResponseEntity<String> getGithubUser(@RequestParam String code) {
+        DataGithubPostRequest data = new DataGithubPostRequest(code);
+        String response = githubService.consumeApi(Constant.GITHUB_AUTH_ACCESSTOKEN_URL, data);
+        return ResponseEntity.status(200).body(response);
+    }
+
+    @PostMapping("/github")
+    public ResponseEntity<UserEntity> postGitHubUser(@RequestBody GithubUserRequest githubUser){
+        UserEntity user =  authService.githubAuth(githubUser);
+        if (user == null) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Something goes Wrong");
+        return ResponseEntity.status(200).body(user);
     }
 }
