+++ src/main/java/controly/service/AuthService.java
@@ -1,6 +1,7 @@
 package controly.service;
 
 
+import controly.dto.GithubUserRequest;
 import controly.entities.UserEntity;
 import controly.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +9,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
+
+import java.util.Optional;
 
 @Service
 public class AuthService {
@@ -23,5 +26,15 @@ public class AuthService {
             throw new ResponseStatusException(HttpStatus.UNAUTHORIZED,"Senha incorreta");
         return ResponseEntity.status(200).body(user);
 
+    }
+
+    public UserEntity githubAuth(GithubUserRequest githubUser) {
+        Optional<UserEntity> userEntityOptional = userRepository.findByIdGithub(githubUser.getIdGithub());
+        if (userEntityOptional.isPresent()){
+            return userEntityOptional.get();
+        }
+        UserEntity newUser = githubUser.convert();
+        userRepository.save(newUser);
+        return newUser;
     }
 }
